Also modify article-categories in MECA post-publication workflow.
