+++ tests/provider/test_cleaner.py
@@ -1416,6 +1416,119 @@ class TestModifyVolume(unittest.TestCase):
         self.assertEqual(xml_string, expected)
 
 
+class TestModifyArticleCategories(unittest.TestCase):
+    "tests for modify_article_categories()"
+
+    def test_modify_article_categories(self):
+        "test removing and adding subject tags"
+        xml_root = ElementTree.fromstring(
+            "<article>"
+            "<front>"
+            "<article-meta>"
+            "<article-categories>"
+            '<subj-group subj-group-type="heading">'
+            "<subject>Research Article</subject>"
+            "</subj-group>"
+            "</article-categories>"
+            "</article-meta>"
+            "</front>"
+            "</article>"
+        )
+        display_channel = "Research Article"
+        article_categories = ["Epidemiology and Global Health"]
+        expected = (
+            "<article>"
+            "<front>"
+            "<article-meta>"
+            "<article-categories>"
+            '<subj-group subj-group-type="display-channel">'
+            "<subject>Research Article</subject>"
+            "</subj-group>"
+            '<subj-group subj-group-type="heading">'
+            "<subject>Epidemiology and Global Health</subject>"
+            "</subj-group>"
+            "</article-categories>"
+            "</article-meta>"
+            "</front>"
+            "</article>"
+        )
+        # invoke
+        cleaner.modify_article_categories(xml_root, display_channel, article_categories)
+        # assert
+        xml_string = ElementTree.tostring(xml_root).decode("utf-8")
+        self.assertEqual(xml_string, expected)
+
+    def test_add_article_categories(self):
+        "test adding article-categories tag"
+        xml_root = ElementTree.fromstring(
+            "<article>"
+            "<front>"
+            "<article-meta>"
+            '<article-id pub-id-type="doi">10.7554/eLife.95901</article-id>'
+            '<article-id pub-id-type="doi" specific-use="version">'
+            "10.7554/eLife.95901.1"
+            "</article-id>"
+            "<title-group>"
+            "<article-title>Title</article-title>"
+            "</title-group>"
+            "</article-meta>"
+            "</front>"
+            "</article>"
+        )
+        display_channel = "Research Article"
+        article_categories = ["Epidemiology and Global Health"]
+        expected = (
+            "<article>"
+            "<front>"
+            "<article-meta>"
+            '<article-id pub-id-type="doi">10.7554/eLife.95901</article-id>'
+            '<article-id pub-id-type="doi" specific-use="version">'
+            "10.7554/eLife.95901.1"
+            "</article-id>"
+            "<article-categories>"
+            '<subj-group subj-group-type="display-channel">'
+            "<subject>Research Article</subject>"
+            "</subj-group>"
+            '<subj-group subj-group-type="heading">'
+            "<subject>Epidemiology and Global Health</subject>"
+            "</subj-group>"
+            "</article-categories>"
+            "<title-group>"
+            "<article-title>Title</article-title>"
+            "</title-group>"
+            "</article-meta>"
+            "</front>"
+            "</article>"
+        )
+        # invoke
+        cleaner.modify_article_categories(xml_root, display_channel, article_categories)
+        # assert
+        xml_string = ElementTree.tostring(xml_root).decode("utf-8")
+        self.assertEqual(xml_string, expected)
+
+    def test_no_article_categories(self):
+        "test no data will remove the article-categories tag"
+        xml_root = ElementTree.fromstring(
+            "<article>"
+            "<front>"
+            "<article-meta>"
+            "<article-categories>"
+            '<subj-group subj-group-type="heading">'
+            "<subject>Research Article</subject>"
+            "</subj-group>"
+            "</article-categories>"
+            "</article-meta>"
+            "</front>"
+            "</article>"
+        )
+        expected = "<article><front><article-meta /></front></article>"
+        # invoke
+        cleaner.modify_article_categories(xml_root)
+        # assert
+        xml_string = ElementTree.tostring(xml_root).decode("utf-8")
+        self.assertEqual(xml_string, expected)
+
+
 class TestModifyPermissions(unittest.TestCase):
     "tests for modify_permissions()"
